1 task/consumer/requirements.txt
pika
psycopg2-binary